Move app to src
