+++ M2/Back-end/pet_shop/src/db/sell_module/db_receipt_manager.py
@@ -1,7 +1,7 @@
 from sqlalchemy import select, insert, update, and_, Table
-from utils.api_exception import APIException
+from src.utils.api_exception import APIException
 from datetime import datetime
-from db.utils_db.helpers import (
+from src.db.utils_db.helpers import (
     _verify_user_own_cart,
     _filter_locals)
