+++ M2/Back-end/pet_shop/src/db/user_module/db_payment_manager.py
@@ -1,6 +1,6 @@
 from sqlalchemy import insert, select, delete, update, and_
-from db.utils_db.helpers import _filter_locals, _filter_values, _verify_user_own_payment
-from utils.api_exception import APIException
+from src.db.utils_db.helpers import _filter_locals, _filter_values, _verify_user_own_payment
+from src.utils.api_exception import APIException
 
 
 class DbPaymentManager:
