+++ M2/Back-end/pet_shop/src/utils/api_exception.py
